+++ etl4/run_load.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+echo '-- Running FIXTURES --'
+python3 fixtures.py
+echo '-- Running LOAD --'
+python3 load.py
