+++ load.sh
@@ -1,3 +1,16 @@
-#!/bin/bash
-echo '-- Running LOAD --'
-python3 load/app.py
+##!/bin/bash
+#export DB_USER=casrec
+#export DB_PASSWORD=casrec # pragma: allowlist secret
+#export DB_NAME=casrecmigration
+#export DB_HOST=casrec_db
+#export DB_PORT=5432
+#export SIRIUS_DB_USER=api
+#export SIRIUS_DB_PASSWORD=api # pragma: allowlist secret
+#export SIRIUS_DB_NAME=api
+#export SIRIUS_DB_HOST=postgres-sirius
+#export SIRIUS_DB_PORT=5432
+#export ENVIRONMENT=local
+#
+#
+#echo '-- Running LOAD --'
+#python3 load/app.py
